Add variables and predicate for Exp without them
We well need to save functions in context in order to build a good
interpreter. Therefore, we need to access the context from inside the
expressions.

We also need to evaluate only the expressions that are free from
dependency on context. We shall later prove decidability of having or
lacking external bindings and request the proofs that an expression
doesn't access context before evaluation.